backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
pyaudio
numpy>=2.1.0
asyncio-mqtt
pinecone
pandas>=2.2.0

